requirements.txt
flask
PyGithub
PyYAML